Move SmellWarning to top level.
