+++ lib/reek/smell_detectors/uncommunicative_method_name.rb
@@ -1,6 +1,5 @@
 # frozen_string_literal: true
 require_relative 'base_detector'
-require_relative 'smell_warning'
 
 module Reek
   module SmellDetectors
