+++ lib/reek/smell_detectors/utility_function.rb
@@ -1,7 +1,6 @@
 # frozen_string_literal: true
 require_relative '../ast/reference_collector'
 require_relative 'base_detector'
-require_relative 'smell_warning'
 
 module Reek
   module SmellDetectors
